Use uint64 instead of uint32

+++ things/api/grpc/client.go
@@ -40,14 +40,14 @@ func NewClient(conn *grpc.ClientConn) mainflux.ThingsServiceClient {
 }
 
 func (client grpcClient) CanAccess(ctx context.Context, req *mainflux.AccessReq, _ ...grpc.CallOption) (*mainflux.ThingID, error) {
-	ar := accessReq{thingKey: req.GetToken(), chanID: uint(req.GetChanID())}
+	ar := accessReq{thingKey: req.GetToken(), chanID: req.GetChanID()}
 	res, err := client.canAccess(ctx, ar)
 	if err != nil {
 		return nil, err
 	}
 
 	ir := res.(identityRes)
-	return &mainflux.ThingID{Value: uint32(ir.id)}, ir.err
+	return &mainflux.ThingID{Value: ir.id}, ir.err
 }
 
 func (client grpcClient) Identify(ctx context.Context, req *mainflux.Token, _ ...grpc.CallOption) (*mainflux.ThingID, error) {
@@ -57,12 +57,12 @@ func (client grpcClient) Identify(ctx context.Context, req *mainflux.Token, _ ..
 	}
 
 	ir := res.(identityRes)
-	return &mainflux.ThingID{Value: uint32(ir.id)}, ir.err
+	return &mainflux.ThingID{Value: ir.id}, ir.err
 }
 
 func encodeCanAccessRequest(_ context.Context, grpcReq interface{}) (interface{}, error) {
 	req := grpcReq.(accessReq)
-	return &mainflux.AccessReq{Token: req.thingKey, ChanID: uint32(req.chanID)}, nil
+	return &mainflux.AccessReq{Token: req.thingKey, ChanID: req.chanID}, nil
 }
 
 func encodeIdentifyRequest(_ context.Context, grpcReq interface{}) (interface{}, error) {
@@ -72,5 +72,5 @@ func encodeIdentifyRequest(_ context.Context, grpcReq interface{}) (interface{},
 
 func decodeIdentityResponse(_ context.Context, grpcRes interface{}) (interface{}, error) {
 	res := grpcRes.(*mainflux.ThingID)
-	return identityRes{uint(res.GetValue()), nil}, nil
+	return identityRes{res.GetValue(), nil}, nil
 }
