+++ things/api/grpc/responses.go
@@ -1,6 +1,6 @@
 package grpc
 
 type identityRes struct {
-	id  uint
+	id  uint64
 	err error
 }
